pyproject.toml
[tool.poetry]
name = "boostdx"
version = "0.1.0"
description = "Poniendo en practica los conocimientos del libro Boost your Django DX"
authors = ["Maynor Pineda"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]
ipython = "^8.3.0"
pywatchman = "^1.4.1"
rich = "^12.4.1"
black = "^22.3.0"
django-debug-toolbar = "^3.4.0"
django-browser-reload = "^1.3.0"

[tool.black]
target-version = ['py38']
line-lenth = 79

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
